pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyopenclaw"
version = "0.2.0"
description = "A self-hosted, persistent AI agent gateway."
authors = [{name = "PyOpenClaw Team"}]
dependencies = [
    "aiosqlite>=0.20",
    "pydantic>=2.0",
    "openai>=1.0",
    "aiohttp>=3.9",
    "websockets>=12.0",
    "structlog>=24.0",
    "watchfiles>=0.21",
    "sentence-transformers>=2.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "coverage[toml]>=7.0",
    "respx>=0.21",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
addopts = """
  --cov=src/pyopenclaw
  --cov-report=term-missing
  --cov-report=html:tests/coverage/html
  --cov-report=xml:tests/coverage/coverage.xml
"""
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/pyopenclaw"]
branch = true
omit = [
    "*/__init__.py",
    "src/pyopenclaw/cli/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
